// apps/studio/src/protocol/types.ts
export interface DeviceInfo {
  product: string;
  hardware: string;
  serial: string;
  protocol: number;
  p4Firmware: string;
  cameraFirmware: [string, string, string, string];
  sensors: [string, string, string, string];
  sdPresent: boolean;
  sdFreeMB: number;
  activeMode: ShootMode;
  activeRecipe: string;
}
// ...
export type ShootMode = 'wiggle' | 'quad';